+++ NEWS.md
@@ -1,5 +1,10 @@
 # SpectroPipeR 0.4.4
 
+* feature: integrate directLFQ for label-free quantification, using the directLFQ python package [MannLabs/directlfq](https://github.com/MannLabs/directlfq) 
+  (Ammar, C., Schessner, J. P., Willems, S., Michaelis, A. C. & Mann, M. 
+  "Accurate label-free quantification by directLFQ to compare unlimited numbers of proteomes", 2023, doi:10.1101/2023.02.17.528962), 
+  adapted via the code of [RdirectLFQ](https://github.com/mildpiggy/Rdirectlfq) package (author: Zhenhuan Feng) with modifications.
+  
 # SpectroPipeR 0.4.3
 
 * feature: add automatic missing value imputation by halve all values within the lowest 1% quantile distribution of ion quantities to create a new dataset used for imputing missing values.